L'action de récupération de logo passe par la fonction idoine (évite de dupliquer du code)
Cette fonction retourne le nouvel id_document lié à la vignette

Pour les documents dont le son est en codé via le codes vorbis ou libvorbis, on utilise un niveau de qualité plutôt qu'un bit rate (Vorbis étant toujours en VBR)

Le cron ne fonctionne que si le chemin vers ffmpeg n'est pas cassé (méta)

Si un document original n'existe plus ses occurrences sont enlevées de la file d'attente via le cron

Amélioration du script d'encodage prenant en compte la qualité des fichiers utilisant vorbis

On augmente le numéro de version du plugin 


git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@38898 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ spipmotion/inc/encodage.php
@@ -23,8 +23,15 @@ function inc_encodage_dist($source,$attente){
  */
 function encodage($source,$doc_attente){
 	/**
+	 * Si le chemin vers le binaire FFMpeg n'existe pas,
+	 * la configuration du plugin crée une meta spipmotion_casse
+	 */
+	if($GLOBALS['meta']['spipmotion_casse'] == 'oui')
+		return;
+
+	/**
 	 * On change le statut d'encodage à en_cours pour
-	 * - changer les messages
+	 * - changer les messages sur le site (ce media est en cours d'encodage par exemple)
 	 * - indiquer si nécessaire le statut
 	 */
 	sql_updateq("spip_spipmotion_attentes",array('encode'=>'en_cours'),"id_spipmotion_attente=".intval($doc_attente));
@@ -33,23 +40,32 @@ function encodage($source,$doc_attente){
 	$extension_attente = $attente['extension'];
 	$type_doc = $attente['objet'];
 	$id_objet = $attente['id_objet'];
-	spip_log($attente,"spipmotion");
 
 	include_spip('inc/documents');
 	$chemin = get_spip_doc($source['fichier']);
 	spip_log("encodage de $chemin","spipmotion");
 
 	$fichier = basename($source['fichier']);
+	/**
+	 * Génération des noms temporaires et finaux
+	 * - Le nom du dossier temporaire (tmp/spipmotion)
+	 * - Le nom du fichier final (nom_du_fichier-encoded.ext)
+	 * - Le nom du fichier temporaire durant l'encodage
+	 * - Le nom du fichier de log généré pour chaque fichier
+	 */
 	$query = "$fichier-$extension_attente-".date('Y_m_d_H-i-s');
 	$dossier = _DIR_TMP.'spipmotion/';
 	$fichier_final = substr($fichier,0,-(strlen($source['extension'])+1)).'-encoded.'.$extension_attente;
-
-	if(!is_dir($dossier)){
-		sous_repertoire(_DIR_TMP,'spipmotion');
-	}
 	$fichier_temp = "$dossier$query.$extension_attente";
 	$fichier_log = "$dossier$query.log";
 	spip_log("le nom temporaire durant l'encodage est $fichier_temp","spipmotion");
+
+	/**
+	 * On crée le dossier temporaire s'il n'existe pas
+	 */
+	if(!is_dir($dossier)){
+		sous_repertoire(_DIR_TMP,'spipmotion');
+	}
 
 	/**
 	 * $texte est le contenu du fichier de preset que l'on passe à la commande
@@ -69,33 +85,48 @@ function encodage($source,$doc_attente){
 	/**
 	 * Correction des bitrates audios si nécessaires
 	 */
-	if(intval($source['audiobitrate']) && (intval($source['audiobitrate']) < lire_config("spipmotion/bitrate_audio_$extension_attente","64"))){
-		$audiobitrates = array('32','64','96','128','192','256');
-		if(!in_array($source['audiobitrate'],$audiobitrates)){
-			$bitrate_final = '32';
-			foreach($audiobitrates as $bitrate){
-				if($source['audiobitrate'] > $bitrate){
-					$bitrate_final = $bitrate;
-				}
-			}
-			$abitrate = $bitrate_final;
-		}else{
-			$abitrate = $source['audiobitrate'];
-		}
+	if(in_array(lire_config("spipmotion/acodec_$extension_attente",''),array('vorbis','libvorbis'))){
+		$qualite = lire_config("spipmotion/qualite_audio_$extension_attente",'4');
+		spip_log('qualite = '.$qualite,'spipmotion');
+		/**
+		 * ffmpeg a besoin d'une valeur entre -10 et 100 alors que ffmpeg2theora respecte libvobis
+		 * de -1 à 10
+		 */
+		$texte .= "aq=".$qualite."\n";
+		$audiobitrate_ffmpeg2theora = "--audioquality $qualite";
+		$audiobitrate_ffmpeg = "--audioquality ".$qualite;
 	}else{
-		$abitrate = lire_config("spipmotion/bitrate_audio_$extension_attente","64");
+		if(intval($source['audiobitrate']) && (intval($source['audiobitrate']) < lire_config("spipmotion/bitrate_audio_$extension_attente","64"))){
+			$audiobitrates = array('32','64','96','128','192','256');
+			if(!in_array($source['audiobitrate'],$audiobitrates)){
+				$bitrate_final = '32';
+				foreach($audiobitrates as $bitrate){
+					if($source['audiobitrate'] > $bitrate){
+						$bitrate_final = $bitrate;
+					}
+				}
+				$abitrate = $bitrate_final;
+			}else{
+				$abitrate = $source['audiobitrate'];
+			}
+		}else{
+			$abitrate = lire_config("spipmotion/bitrate_audio_$extension_attente","64");
+		}
+		$texte .= "ab=".$abitrate."000\n";
+		$audiobitrate_ffmpeg = $audiobitrate_ffmpeg2theora = "--audiobitrate ".$abitrate;
 	}
-
-	$texte .= "ab=".$abitrate."000\n";
-	$audiobitrate = "--audiobitrate ".$abitrate;
 
 	/**
 	 * Vérification des samplerates
 	 */
 	if(intval($source['audiosamplerate']) && (intval($source['audiosamplerate']) < lire_config("spipmotion/frequence_audio_$extension_attente","22050"))){
-		$audiosamplerates = array('11025','16000','22050','44100','48000');
+		$audiosamplerates = array('8000','11025','16000','22050','24000','32000','44100','48000');
+		if(lire_config("spipmotion/acodec_$extension_attente",'') == 'libmp3lame'){
+			unset($audiosamplerates['11025']);
+			unset($audiosamplerates['8000']);
+		}
 		if(!in_array($source['audiosamplerate'],$audiosamplerates)){
-			$audiosamplerate_final = '11025';
+			$audiosamplerate_final = min($audiosamplerates);
 			foreach($audiosamplerates as $samplerate){
 				if($source['audiosamplerate'] >= $samplerate){
 					$audiosamplerate_final = $samplerate;
@@ -116,15 +147,16 @@ function encodage($source,$doc_attente){
 	 * le format choisi est vorbis (l'encodeur vorbis de ffmpeg ne gère pas le mono)
 	 */
 	if(($source['audiochannels'] > 2) OR (in_array($extension_attente,array('ogg','ogv','oga')) && ($source['audiochannels'] < 2))){
-		spip_log('on passe en deux canaux','test');
+		spip_log('on passe en deux canaux','spipmotion');
 		$audiochannels = 2;
 	}else{
+		spip_log($source['audiochannels'].' canaux','spipmotion');
 		$audiochannels = $source['audiochannels'];
 	}
 
 	if(intval($audiochannels) >= 1){
 		$texte .= "ac=$audiochannels\n";
-		$audiochannels = "--ac $audiochannels";
+		$audiochannels_ffmpeg = "--ac $audiochannels";
 	}
 
 	/**
@@ -134,11 +166,12 @@ function encodage($source,$doc_attente){
 		/**
 		 * Encodage du son
 		 */
-		$encodage = find_in_path('script_bash/spipmotion.sh').' --e '.$chemin.' --s '.$fichier_temp.' '.$audiobitrate.' '.$audiofreq.' '.$audiochannels.' --p '.lire_config("spipmotion/chemin","/usr/local/bin/ffmpeg").' &> '.$fichier_log;
+		$acodec = lire_config("spipmotion/acodec_$extension_attente") ? "--acodec ".lire_config("spipmotion/acodec_$extension_attente") :'';
+		$encodage = find_in_path('script_bash/spipmotion.sh').' --e '.$chemin.' --s '.$fichier_temp.' '.$acodec.' '.$audiobitrate_ffmpeg.' '.$audiofreq.' '.$audiochannels_ffmpeg.' --p '.lire_config("spipmotion/chemin","/usr/local/bin/ffmpeg").' &> '.$fichier_log;
 		spip_log("$encodage",'spipmotion');
-		$lancement_encodage = exec($encodage,$retour);
+		$lancement_encodage = exec($encodage,$retour,$retour_int);
+		spip_log($retour_int,'spipmotion');
 		spip_log($retour,'spipmotion');
-		spip_log("l'encodage est terminé",'spipmotion');
 		if(filesize($fichier_temp) > 100){
 			$encodage_ok = true;
 		}else{
@@ -241,14 +274,13 @@ function encodage($source,$doc_attente){
 		 * sinon on utilise notre script pour ffmpeg
 		 */
 		if((lire_config("spipmotion/encodeur_$extension_attente",'') == 'ffmpeg2theora') && (lire_config('spipmotion_ffmpeg2theora/version') > 0)){
-			$encodage = "ffmpeg2theora $chemin -v 8 -V $vbitrate -A $abitrate -H $samplerate --max_size ".$width_finale."x".$height_finale." -F $fps_num --optimize --nice 9 -o $fichier_temp &> $fichier_log";
+			$encodage = "ffmpeg2theora $chemin -v 8 -V $vbitrate $audiobitrate_ffmpeg2theora -H $samplerate -c $audiochannels --max_size ".$width_finale."x".$height_finale." -F $fps_num --optimize --nice 9 -o $fichier_temp &> $fichier_log";
 		}else{
-			$encodage = find_in_path('script_bash/spipmotion.sh')." $audiofreq $video_size --e $chemin $acodec $vcodec $fps $audiobitrate $audiochannels $bitrate $vpre $vpre2 --s $fichier_temp --fpre $fichier_texte --p ".lire_config("spipmotion/chemin","/usr/local/bin/ffmpeg")." &> $fichier_log";
+			$encodage = find_in_path('script_bash/spipmotion.sh')." $audiofreq $video_size --e $chemin $acodec $vcodec $fps $audiobitrate $audiochannels_ffmpeg $bitrate $vpre $vpre2 --s $fichier_temp --fpre $fichier_texte --p ".lire_config("spipmotion/chemin","/usr/local/bin/ffmpeg")." &> $fichier_log";
 		}
 		spip_log($encodage,'spipmotion');
-		$lancement_encodage = exec($encodage,$retour);
-
-		spip_log($retour,'spipmotion');
+		$lancement_encodage = exec($encodage,$retour,$retour_int);
+		spip_log($retour_int,'spipmotion');
 
 		if(filesize($fichier_temp) > 100){
 			$encodage_ok = true;
@@ -265,8 +297,8 @@ function encodage($source,$doc_attente){
 			 * Inscrire les metadatas dans la video finale
 			 */
 			$metadatas_flv = 'flvtool2 -Ux '.$fichier_temp;
-			exec($metadatas_flv,$retour);
-			spip_log($retour,'spipmotion');
+			exec($metadatas_flv,$retour,$retour_int);
+			spip_log($retour_int,'spipmotion');
 		}
 	}
 
@@ -279,13 +311,17 @@ function encodage($source,$doc_attente){
 		$mode = 'document';
 		$invalider = true;
 
-		sql_updateq("spip_spipmotion_attentes",array('encode'=>'oui'),"id_spipmotion_attente=".intval($doc_attente));
-
 		$ajouter_documents = charger_fonction('ajouter_documents', 'inc');
 		$x = $ajouter_documents($fichier_temp, $fichier_final, $type_doc, $id_objet, $mode, '', $actif,'','','');
-		unlink($fichier_temp);
 
-		sql_updateq("spip_documents",array('id_orig'=>$attente['id_document']),'id_document='.intval($x));
+		if(intval($x) > 1){
+			unlink($fichier_temp);
+			sql_updateq("spip_spipmotion_attentes",array('encode'=>'oui'),"id_spipmotion_attente=".intval($doc_attente));
+			sql_updateq("spip_documents",array('id_orig'=>$attente['id_document']),'id_document='.intval($x));
+		}else{
+			sql_updateq("spip_spipmotion_attentes",array('encode'=>'non'),"id_spipmotion_attente=".intval($doc_attente));
+			spip_log('Il y a une erreur, le fichier n est pas copié','spipmotion');
+		}
 
 		/**
 		 * Invalidation du cache
